third_party/WebKit/LayoutTests/fast/events/wheel/mainthread-touchpad-fling-latching.html

<!DOCTYPE HTML>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<style>

  ::-webkit-scrollbar {
    display: none;
  }
  body {
    margin: 0px;
    height: 1000px;
    width: 1000px;
  }
  #parentDiv {
    background-color: #FF7F7F;
    height: 600px;
    width: 600px;
    overflow: scroll;
  }
  #content1 {
    height: 700px;
    width: 700px;
  }
  #childDiv {
    background-color: #84BE6A;
    height: 500px;
    width: 500px;
    overflow: scroll;
  }
  #content2 {
    height: 600px;
    width: 600px;
  }
</style>

<div id="parentDiv">
  <div id="content1">
    <div id="childDiv">
      <div id="content2">
      </div>
    </div>
  </div>
</div>

<script>

var parentDiv = document.getElementById('parentDiv');
var childDiv = document.getElementById('childDiv');
var rect = childDiv.getBoundingClientRect();

setup({ explicit_done: true });
test(() => {
  // Start scrolling on the child div.
  eventSender.gestureScrollBegin("touchpad",
                                 (rect.left + rect.right) / 2,
                                 (rect.top + rect.bottom) / 2);
  eventSender.gestureScrollUpdate("touchpad", 0, -70);
  waitForChildScrollAndCheck();
    
}, "Main thread fling on child div should latch to the div.");

var lastScrollOffset = childDiv.scrollTop;
var rafCount = 0;

// Wait for the child div to fully consume scroll delta, then start a fling.
function waitForChildScrollAndCheck() {
  if (childDiv.scrollTop == 70) {
    lastScrollOffset = childDiv.scrollTop;
    rafCount = 0;
    sendFlingStart();
  } else if (childDiv.scrollTop != lastScrollOffset) {
    lastScrollOffset = childDiv.scrollTop;
    rafCount = 0;
    requestAnimationFrame(waitForChildScrollAndCheck);
  } else if (rafCount < 10) {
    requestAnimationFrame(waitForChildScrollAndCheck);
    rafCount++;
  } else {
    // Fail the test since for the last 10 rafs childDiv.scrollTop hasn't
    // changed and the scroll delta is not fully consumed.
    assert_equals(childDiv.scrollTop, 70,
        "The child div should fully comsume scroll delta.");
  }
}

function sendFlingStart() {
  // Generate a fling on the child div.
  eventSender.gestureFlingStart((rect.left + rect.right) / 2,
                                (rect.top + rect.bottom) / 2,
                                0, -10000, "touchpad");
  childFlingCheck();
}

// Wait till the child div is fully scroll, then check to see if the fling
// propagates to the parent div or not.
function childFlingCheck() {
  if (childDiv.scrollTop == childDiv.scrollHeight - childDiv.clientHeight) {
    rafCount = 0;
    flingLatchingCheck();
  } else if (childDiv.scrollTop != lastScrollOffset) {
    lastScrollOffset = childDiv.scrollTop;
    rafCount = 0;
    requestAnimationFrame(childFlingCheck);
  } else if (rafCount < 10) {
    requestAnimationFrame(childFlingCheck);
    rafCount++;
  } else {
    // Fail the test since for the last 10 rafs childDiv.scrollTop hasn't
    // changed and the child div is not fully scrolled.
    assert_equals(childDiv.scrollTop,
                  childDiv.scrollHeight - childDiv.clientHeight,
                  "The child div should fully scroll.");
  }
}

function flingLatchingCheck() {
  if (parentDiv.scrollTop != 0) {
    // Fail the test since the fling is not supposed to propagate to the parent
    // div.
    assert_equals(parentDiv.scrollTop, 0,
        "The parent div is not supposed to scroll.");
  } else if (rafCount == 10) {
    // parentDiv.scrollTop remains 0 for 10 rafs. Fling is latched to the child
    // div.
    done();
  } else {
    rafCount++;
    requestAnimationFrame(flingLatchingCheck);
  }
}

</script>
